backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
supabase==2.10.0

# Redis
redis==5.2.1

# Auth & Security
python-jose[cryptography]==3.3.0
pyjwt[crypto]==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Validation & Utils
pydantic==2.10.3
pydantic-settings==2.6.1

# WebSockets (for notifications)
python-socketio==5.11.4
python-engineio==4.9.1

# HTTP Client
httpx==0.27.2
